src/main/java/ru/job4j/ood/srp/report/ReportForHR.java
package ru.job4j.ood.srp.report;

import ru.job4j.ood.srp.formatter.DateTimeParser;
import ru.job4j.ood.srp.model.Employee;
import ru.job4j.ood.srp.store.Store;

import java.util.Calendar;
import java.util.Collections;
import java.util.Comparator;
import java.util.List;
import java.util.function.Predicate;

public class ReportForHR implements Report {
    private final Store store;
    private final DateTimeParser<Calendar> dateTimeParser;

    public ReportForHR(Store store, DateTimeParser<Calendar> dateTimeParser) {
        this.store = store;
        this.dateTimeParser = dateTimeParser;
    }
    class ItemDescBySalary implements Comparator<Employee> {
        public int compare(Employee left, Employee right) {
            return (int) (right.getSalary() * 100 - left.getSalary() * 100);
        }
    }
    @Override
    public String generate(Predicate<Employee> filter) {
        StringBuilder text = new StringBuilder();
        text.append("Name; Hired; Fired; Salary;")
                .append(System.lineSeparator());
        List<Employee> emp = store.findBy(filter);
        emp.sort(new ItemDescBySalary());
        for (Employee employee : emp) {
            text.append(employee.getName()).append(" ")
                    .append(employee.getSalary())
                    .append(System.lineSeparator());
        }
        return text.toString();
    }
}
